ci: compliance: split ClangFormatCheck into compliance/ directory
Move ClangFormatCheck out of check_compliance.py and into
scripts/ci/compliance/clang_format.py.

The new module uses _compliance_context for GIT_TOP and COMMIT_RANGE.
The now-unused imports (platform, unidiff) are removed from
check_compliance.py.
