+++ scripts/ci/check_compliance.py
@@ -9,7 +9,6 @@ import collections
 import json
 import logging
 import os
-import platform
 import re
 import shlex
 import shutil
@@ -40,7 +39,6 @@ except ImportError:
         if dll_dir.is_dir():
             os.environ['PATH'] = str(dll_dir) + os.pathsep + os.environ.get('PATH', '')
     import magic
-import unidiff
 import yaml
 from dotenv import load_dotenv
 from junitparser import Error, Failure, JUnitXml, Skipped, TestCase, TestSuite
@@ -417,58 +415,6 @@ class BoardYmlCheck(ComplianceTest):
 
         for file in path.glob("**/board.yml"):
             self.check_board_file(file, vendor_prefixes)
-
-
-class ClangFormatCheck(ComplianceTest):
-    """
-    Check if clang-format reports any issues
-    """
-
-    name = "ClangFormat"
-    doc = zephyr_doc_detail_builder("/contribute/guidelines.html#clang-format")
-
-    def _process_patch_error(self, file: str, patch: unidiff.PatchedFile):
-        for hunk in patch:
-            # Strip the before and after context
-            before = next(i for i, v in enumerate(hunk) if str(v).startswith(('-', '+')))
-            after = next(i for i, v in enumerate(reversed(hunk)) if str(v).startswith(('-', '+')))
-            msg = "".join([str(line) for line in hunk[before : -after or None]])
-
-            # show the hunk at the last line
-            self.fmtd_failure(
-                "notice",
-                "You may want to run clang-format on this change",
-                file,
-                line=hunk.source_start + hunk.source_length - after,
-                desc=f'\r\n{msg}',
-            )
-
-    def run(self):
-        exe = f"clang-format-diff.{'exe' if platform.system() == 'Windows' else 'py'}"
-
-        for file in get_files():
-            if Path(file).suffix not in ['.c', '.h']:
-                continue
-
-            diff = subprocess.Popen(
-                ('git', 'diff', '-U0', '--no-color', COMMIT_RANGE, '--', file),
-                stdout=subprocess.PIPE,
-                cwd=GIT_TOP,
-            )
-            try:
-                subprocess.run(
-                    (exe, '-p1'),
-                    check=True,
-                    stdin=diff.stdout,
-                    stdout=subprocess.PIPE,
-                    stderr=subprocess.STDOUT,
-                    cwd=GIT_TOP,
-                )
-
-            except subprocess.CalledProcessError as ex:
-                patchset = unidiff.PatchSet.from_string(ex.output, encoding="utf-8")
-                for patch in patchset:
-                    self._process_patch_error(file, patch)
 
 
 class DevicetreeBindingsCheck(ComplianceTest):
